backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles==24.1.0",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asgiref==3.8.1",
    "attrs==25.3.0",
    "backoff==2.2.1",
    "bcrypt==4.3.0",
    "build==1.2.2.post1",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "chromadb==1.0.12",
    "click==8.2.1",
    "colorama==0.4.6",
    "coloredlogs==15.0.1",
    "defusedxml==0.7.1",
    "distro==1.9.0",
    "durationpy==0.10",
    "fastapi==0.115.9",
    "filelock==3.18.0",
    "flatbuffers==25.2.10",
    "fsspec==2025.5.1",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.25.0",
    "google-api-python-client==2.171.0",
    "google-auth==2.40.3",
    "google-auth-httplib2==0.2.0",
    "google-generativeai==0.8.5",
    "googleapis-common-protos==1.70.0",
    "grpcio==1.72.1",
    "grpcio-status==1.71.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "huggingface-hub==0.32.4",
    "humanfriendly==10.0",
    "idna==3.10",
    "importlib-metadata==8.7.0",
    "importlib-resources==6.5.2",
    "jsonschema==4.24.0",
    "jsonschema-specifications==2025.4.1",
    "kubernetes==32.0.1",
    "markdown-it-py==3.0.0",
    "mdurl==0.1.2",
    "mmh3==5.1.0",
    "mpmath==1.3.0",
    "numpy==2.2.6",
    "oauthlib==3.2.2",
    "onnxruntime==1.22.0",
    "opentelemetry-api==1.34.0",
    "opentelemetry-exporter-otlp-proto-common==1.34.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.34.0",
    "opentelemetry-instrumentation==0.55b0",
    "opentelemetry-instrumentation-asgi==0.55b0",
    "opentelemetry-instrumentation-fastapi==0.55b0",
    "opentelemetry-proto==1.34.0",
    "opentelemetry-sdk==1.34.0",
    "opentelemetry-semantic-conventions==0.55b0",
    "opentelemetry-util-http==0.55b0",
    "orjson==3.10.18",
    "overrides==7.7.0",
    "packaging==25.0",
    "posthog==4.4.0",
    "proto-plus==1.26.1",
    "protobuf==5.29.5",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.11.5",
    "pydantic-core==2.33.2",
    "pygments==2.19.1",
    "pyparsing==3.2.3",
    "pypdf2==3.0.1",
    "pypika==0.48.9",
    "pyproject-hooks==1.2.0",
    "pyreadline3==3.5.4",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "referencing==0.36.2",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "rich==14.0.0",
    "rpds-py==0.25.1",
    "rsa==4.9.1",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "starlette==0.45.3",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "tokenizers==0.21.1",
    "tqdm==4.67.1",
    "typer==0.16.0",
    "typing-extensions==4.14.0",
    "typing-inspection==0.4.1",
    "uritemplate==4.2.0",
    "urllib3==2.4.0",
    "uvicorn==0.34.3",
    "watchfiles==1.0.5",
    "websocket-client==1.8.0",
    "websockets==15.0.1",
    "wrapt==1.17.2",
    "youtube-transcript-api==1.0.3",
    "zipp==3.22.0",
]
